chore: Made a separate check on the devices
