+++ my-vue-app/src/utils/screen.js
@@ -1,0 +1,20 @@
+import { ref, onMounted, onBeforeUnmount } from 'vue';
+
+export function useScreenSize() {
+  const isMobile = ref(false);
+
+  const checkScreenSize = () => {
+    isMobile.value = window.innerWidth < 768;
+  };
+
+  onMounted(() => {
+    checkScreenSize();
+    window.addEventListener('resize', checkScreenSize);
+  });
+
+  onBeforeUnmount(() => {
+    window.removeEventListener('resize', checkScreenSize);
+  });
+
+  return { isMobile };
+}
